Improve overtaking AI
Now it's basing its "am I gonna overtake?" on if it's gonna hit the vehicle or get closer than the ideal overtake point (if the forward+rotation vector intersects with the to-be-overtaken forward+rotation vector, or with the to-be-overtaken overtake points line). Then it usually picks the point with the smallest angle, unless that means it needs to cross over the vehicle, in which case it picks the point closest to the track.

Result: Good boi AI with clean-ish overtake lines, good dodging ability and slower but grippier cars can fight back in corner-heavy tracks :D All dynamic and shit.

Needs a heavy cleanup, I probably have more double code here than in any other projects.